Fixed Ambassador Reserve button (#61)
* Add special hour blocks for furniture install (#352)

* Fix low-hanging vulnerabilities (#354)

* Fix conflicting room reservations bug (#355)

* Prevent double booking of the same room

The reservation draft API does not prevent a user from double booking a
room some other user has already booked. While (generally) not possible
via the frontend, this ensures the backend does not permit double booking.

Tests added for various relevant cases.

* Fix failing tests for reservations and warning in organizations

* Fix front-end issue with draft drop-in reservation (#356)

Due to some changes that merged with room reservations, when creating a
drop-in reservation and navigating back (intentionally or not) to home
you would not see the draft reservation anymore because it was filtered
from the home page. This meant a draft reservation was created but could
not be cancelled by user or ambassador. This PR fixes this issue by
making the draft of a drop-in visible on the home page. No changes in
behavior for room reservations.

* Upgrade FastAPI for DDoS security fix to 0.110.0 (#358)

* Profile Page Improvements (#359)

* Created a profile module to contain the new Profile Page and Profile Editor.
* Created a new Profile Page component and associated HTML and CSS files, which includes a button route to the Profile Editor.
* Created a new User Profile Card widget to display a users name, profile picture, email, and pronouns.
* Moved Github Link and Bearer Token to Profile Page and made minor changes to the styling.
* Created a "Profile Actions" card for github link, community agreement, and auth bearer token.

* Fix profile card styling (#361)

* Update policy for furniture install (#362)

* Prevent communal area drop-ins while furniture installation completes (#363)

* Update policy for furniture install

* Block off communal area for OH

* moved design document into doc folder

* updated for all changes

* Make profile page responsive (#366)

* 374 Fix active reservations boundary logic (#375)

* Update filters to show active reservations on ambassador view

* Fix bug when start_idx < 0

* Only consider rooms that are present in map

---------

Co-authored-by: Yuvraj Jain <yjain@unc.edu>

* Reset policies for Phase 2 Launch of XL (#376)

* Fix broken build in prod due to npm memory limit (#377)

* Fix @angular-devkit/build-angular dependency vulnerabilities in production (#378)

* Fix vulnerability with @angular-devkit/build-angular dep

* Try reducing memory to match cloudapps pod limit (#379)

* Increase memory limit for deployment (#381)

* Hot fixing office hours schedule (#382)

* Fix bug with end_idx logic for future reservations (#384)

* Fix bug with end_idx logic for future reservations

The current problem with the code is that at 11 pm, we were
not able to see future reservations in red for tomorrow. This was
actually a red herring that led me to look at parts of the code
that were not responsible for this.

Then suddenly at 12 am, everything started working as intended.
This was the biggest hint for me in debugging this.

Basically on line 345, I just had to add in a condition to check
that the date I am getting as a param is the same date as today.
Only in that case, should I execute the logic of
end_idx < current_time_idx. However, earlier this was being
executed for all reservations which is why whenever we had
any reservations which had an end time of less than 11 pm
(which most reservations do), our function would just skip over
them. This also aligns with what we saw at 12 am when
everything seemed to work as intended. This is again because
most reservations don't have an ending time before 12 am.

Anyway, took me over 60 minutes to add in these 30 chars to fix
the bug. So I typed an average of 0.5 char per minute... :,)

* Added kind of a test for future reservations

Although this test is time dependent, and
we will need some way of freezing time
to accurately test the efficacy of the code.

* Change Dockerfile to use yarn instead of npm (#393)

* 399 add checkin time to avoid cancellation verbiage (#400)

* Add checkin deadline verbiage

* Add parameter to checkinDeadline

Since users can reserve a reservation up until the end a reservation time period, the checkinDeadline was exceeding the reservation's end time.
To correct this take the minimum of the start time + 10 mins and the end time.

* sprint2 branch

* added new folder (#45)

* Jake Sprint2 Admin Updates (#46)

* added new folder

* admin frontend components added

* Made changes to the CSXL Seat Map UI (#52)

* Disable reserve button (#53)

* Made changes to the CSXL Seat Map UI

* Reserve Seats now disabled when CSXL closed

* Seats grey when reserved (#54)

* Made changes to the CSXL Seat Map UI

* Reserve Seats now disabled when CSXL closed

* Seats on widget turn grey when reserved

* Skanta sprint2 (#58)

* make a seat widget

* changed seat data numbers to match and seat map visible

* minor changes

* Backend Admin Updates (#59)

* added new folder

* admin frontend components added

* autopopulate admin form

* unfinished admin backend work

* trying to make backend progress

* fixing backend to frontend connection for admins

* finishing this

---------

Co-authored-by: Jake Rogers <jmanerr@ad.unc.edu>

* Fixed form UI and made sure all map widgets look the same - got rid of functionality for map widgets that don't need it

* Fixed Ambassador Reserve button

---------

Co-authored-by: Kris Jordan <kris@cs.unc.edu>
Co-authored-by: Jade Keegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: Jake Rogers <jmanerr@ad.unc.edu>
Co-authored-by: aarjav <97541934+aarjavjain2002@users.noreply.github.com>
Co-authored-by: Yuvraj Jain <yjain@unc.edu>
Co-authored-by: Yuvraj Jain <97535654+yuvrajjain2003@users.noreply.github.com>
Co-authored-by: Noah Smith <nsmith2@unc.edu>
Co-authored-by: John Schachte <89204366+JohnSchachte@users.noreply.github.com>
Co-authored-by: upasanal <111466862+upasanal@users.noreply.github.com>
Co-authored-by: Jake Rogers <97641772+Jakerogers1@users.noreply.github.com>
Co-authored-by: Shreeya Kantamsetty <111532242+skanta04@users.noreply.github.com>